[EMBR-13976] Feature: Extract FrameTimingSource from FrameRateMonitor (refactor-only)
Splits the CADisplayLink/targetTimestamp-delta primitive out of
FrameRateMonitor into a standalone FrameTimingSource, leaving
FrameRateMonitor to own only the hang-threshold policy and
HangObserver callbacks. HangCaptureService is unaffected since it
only depends on FrameRateMonitor's existing public surface. No
behavior change.
