+++ Sources/EmbraceCore/Capture/Hang/Watchdog/FrameRateMonitor.swift
@@ -5,25 +5,16 @@
 #if !os(watchOS) && !os(macOS)
 
     import Foundation
-    import QuartzCore
 
     #if !EMBRACE_COCOAPOD_BUILDING_SDK
         import EmbraceCommonInternal
     #endif
 
-    /// Monitors frame rendering timing to detect main-thread hangs.
+    /// Detects main-thread hangs from `FrameTimingSource`'s frame-delay reports.
     ///
-    /// `FrameRateMonitor` uses `CADisplayLink` to observe frame delivery on the
-    /// main thread. On each frame, it compares the actual delivery time against
-    /// the system's own committed schedule (`targetTimestamp` from the previous
-    /// tick). A delay exceeding `threshold` is reported as a completed hang to
-    /// the `hangObserver`.
-    ///
-    /// This approach is dynamic-rate–safe: using `targetTimestamp` rather than a
-    /// fixed frame duration means ProMotion, Low Power Mode, and
-    /// `preferredFrameRateRange` transitions never produce false positives.
-    /// `CADisplayLink` also pauses automatically in the background, so suspend
-    /// gaps are excluded without any extra bookkeeping.
+    /// On each frame delay report, a delay exceeding `threshold` is reported as a completed hang
+    /// to the `hangObserver`. Because `FrameTimingSource` only reports a delay once the hang is
+    /// already over, the hang is reported retroactively.
     final class FrameRateMonitor {
 
         /// Apple's own definition of a hang (≈ 250 ms).
@@ -44,62 +35,18 @@
         /// Must be called on the main thread.
         init(threshold: TimeInterval = FrameRateMonitor.defaultAppleHangThreshold) {
             self.threshold = threshold
-            self.proxy = DisplayLinkProxy()
+            self.timingSource = FrameTimingSource()
 
-            let link = CADisplayLink(target: proxy, selector: #selector(DisplayLinkProxy.tick(_:)))
-            link.add(to: .main, forMode: .common)
-            self.displayLink = link
-
-            proxy.monitor = self
-
-            NotificationCenter.default.addObserver(
-                self,
-                selector: #selector(resetOnForeground),
-                name: FrameRateMonitor.willEnterForegroundNotification,
-                object: nil
-            )
-        }
-
-        deinit {
-            NotificationCenter.default.removeObserver(self)
-            displayLink?.invalidate()
+            timingSource.onTick = { [weak self] delay in
+                self?.handle(delay: delay)
+            }
         }
 
         // MARK: - Private
 
-        private let proxy: DisplayLinkProxy
-        private var displayLink: CADisplayLink?
+        private let timingSource: FrameTimingSource
 
-        /// The previous frame's `targetTimestamp` — the system's promise of when
-        /// the current frame would fire.
-        private var previousTickExpectedTimestamp: CFTimeInterval?
-
-        /// Raw notification name to avoid a direct UIKit dependency.
-        private static let willEnterForegroundNotification =
-            Notification.Name("UIApplicationWillEnterForegroundNotification")
-
-        /// Resets state on foreground so the first tick after a background/foreground
-        /// transition is not misreported as a hang.
-        @objc private func resetOnForeground() {
-            previousTickExpectedTimestamp = nil
-        }
-    }
-
-    // MARK: - Frame tick
-
-    extension FrameRateMonitor {
-
-        fileprivate func tick(_ currentTick: CADisplayLink) {
-            defer {
-                previousTickExpectedTimestamp = currentTick.targetTimestamp
-            }
-
-            guard let expectedTimestamp = previousTickExpectedTimestamp else {
-                // First tick: arm for the next frame, nothing to compare yet.
-                return
-            }
-
-            let delay = currentTick.timestamp - expectedTimestamp
+        private func handle(delay: TimeInterval) {
             guard delay > threshold else { return }
 
             // The main thread was blocked beyond `threshold`.
@@ -113,21 +60,6 @@
 
             hangObserver?.hangStarted(at: previousTickDate, duration: delay)
             hangObserver?.hangEnded(at: now, duration: delay)
-        }
-    }
-
-    // MARK: - Weak proxy
-
-    extension FrameRateMonitor {
-
-        /// Holds a weak reference to `FrameRateMonitor` to break the retain cycle
-        /// that `CADisplayLink` would otherwise form with its target.
-        private final class DisplayLinkProxy: NSObject {
-            weak var monitor: FrameRateMonitor?
-
-            @objc func tick(_ link: CADisplayLink) {
-                monitor?.tick(link)
-            }
         }
     }
